more gpu set  up

+++ newton-4.00/sdk/dBrain/gpu/ndBrainGpuBuffer.cpp
@@ -22,7 +22,7 @@ ndInt64 ndBrainGpuBuffer::m_memoryUsed = 0;
 ndBrainGpuBuffer::ndBrainGpuBuffer(ndBrainGpuContext* const context, ndInt64 sizeInByte, ndStorageBufferType bufferTypeFlags, ndDeviceBufferType deviceType)
 	:ndContainersFreeListAlloc<ndBrainGpuBuffer>()
 	,m_context(context)
-	,m_sizeInBytes(sizeInByte)
+	,m_sizeInBytes(size_t(sizeInByte))
 	,m_bufferType(bufferTypeFlags)
 	,m_deviceBufferType(deviceType)
 {
